# openapi.yaml
openapi: 3.0.3
info:
  title: WinOJ API
  description: |
    Windows 在线评测系统 API。支持多语言代码评测、安全沙箱执行、自定义计分脚本、比赛管理、文章系统、Rating 机制、AI 代码安全审查等功能。

    ## 认证
    大部分接口需携带 JWT Token：`Authorization: Bearer <access_token>`
    Access Token 有效期 15 分钟，过期后通过 `POST /auth/refresh` 刷新。

    ## 错误响应
    ```json
    { "code": 1|2|3|4|5|6, "reason": "ERR_xxx", "message": "..." }
    ```
    错误码：1=参数错误, 2=状态错误, 3=未找到, 4=限流, 5=未认证, 6=权限不足

    ## 角色层级
    user < teacher < admin < su
  version: 1.6.0
  contact:
    name: WinOJ
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: "错误码: 1=参数错误, 2=状态错误, 3=未找到, 4=限流, 5=未认证, 6=权限不足"
        reason:
          type: string
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nickname:
          type: string
        role:
          type: string
          enum: [user, teacher, admin, su]
        rating:
          type: integer
        signature:
          type: string
        bio:
          type: string
        email:
          type: string
        hide_rating:
          type: integer
          enum: [0, 1]
        preferred_language:
          type: string
        max_file_size:
          type: integer
          description: 单文件大小限制（字节），0=使用系统默认
        max_storage:
          type: integer
          description: 总存储限制（字节），0=使用系统默认
        created_at:
          type: string
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
    SearchParam:
      name: search
      in: query
      schema:
        type: string
  responses:
...
    NotFound:
      description: 未找到
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /users:
    get:
      tags: [用户]
      summary: 用户列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: role
          in: query
          schema:
            type: string
            enum: [user, teacher, admin, su]
      responses:
        '200':
          description: 用户列表
    post:
      tags: [用户]
      summary: 创建用户
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                nickname:
                  type: string
                role:
                  type: string
                  enum: [user, teacher, admin, su]
                  default: user
      responses:
        '201':
          description: 创建成功
  /users/me:
    get:
      tags: [用户]
      summary: 获取当前用户信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [用户]
      summary: 更新个人资料
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                signature:
                  type: string
                  maxLength: 1000
                bio:
                  type: string
                  description: Markdown 格式
                hide_rating:
                  type: boolean
      responses:
        '200':
          description: 更新成功
  /users/me/virtual-contests:
    get:
      tags: [用户]
      summary: 我的虚拟比赛列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 虚拟比赛列表
  /users/{id}/profile:
    get:
      tags: [用户]
      summary: 获取公开用户资料
      description: 无需登录即可查看用户的公开资料（昵称、签名、简介、Rating 等）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 用户资料
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  nickname:
                    type: string
                  role:
                    type: string
                  signature:
                    type: string
                  bio:
                    type: string
                  rating:
                    type: integer
                  created_at:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /users/rating:
    get:
      tags: [用户]
      summary: Rating 排行榜
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 排行榜
  /users/online:
    get:
      tags: [用户]
      summary: 在线用户列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 在线用户
  /users/sudo-login:
    post:
      tags: [用户]
      summary: 免密登录
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: integer
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      tags: [用户]
      summary: 获取用户详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 用户详情
    delete:
      tags: [用户]
      summary: 删除用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
  /users/{id}/role:
    put:
      tags: [用户]
      summary: 修改角色
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [user, teacher, admin, su]
      responses:
        '200':
          description: 修改成功
  /users/{id}/ban:
    post:
      tags: [用户]
      summary: 封禁用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 封禁成功
  /users/{id}/unban:
    post:
      tags: [用户]
      summary: 解封用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 解封成功
  /users/{id}/force-logout:
    post:
      tags: [用户]
      summary: 强制登出
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 强制登出成功
  /users/{id}/reset-password:
    post:
      tags: [用户]
      summary: 重置密码
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                new_password:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: 重置成功
  /users/{id}/rating:
    put:
      tags: [用户]
      summary: 修改 Rating
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
      responses:
        '200':
          description: 修改成功
  /users/{id}/hide-rating:
    put:
      tags: [用户]
      summary: 隐藏/显示 Rating
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hide_rating]
              properties:
                hide_rating:
                  type: integer
                  enum: [0, 1]
      responses:
        '200':
          description: 设置成功
  /users/{id}/submit-lock-exempt:
    put:
      tags: [用户]
      summary: 设置提交锁豁免
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [submit_lock_exempt]
              properties:
                submit_lock_exempt:
                  type: boolean
      responses:
        '200':
          description: 设置成功
  /users/{id}/upload-limits:
    put:
      tags: [用户]
      summary: 设置图床限制
      description: 超级管理员设置用户单文件大小和总存储限制，0 表示使用系统默认（10MB / 2GB）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                max_file_size:
                  type: integer
                  description: 单文件大小限制（字节），0=默认10MB
                max_storage:
                  type: integer
                  description: 总存储限制（字节），0=默认2GB
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  max_file_size:
                    type: integer
                  max_storage:
                    type: integer

  # ===== LANGUAGES =====
...